refactor: split TOC components
