+++ src/components/toc/index.tsx
@@ -1,0 +1,2 @@
+export { TOCMinimap } from './minimap'
+export { TOCPopover } from './popover'
